requirements.txt
 streamlit
 SpeechRecognition
 google-generativeai
pygame
gtts